Move change-limit-modal into subfolder
Place change-limit-modal component files into a dedicated folder
